Resolving remaining conflicts in execution test suite (#171)

+++ test/libsolidity/semanticTests/array/fixed_out_of_bounds_array_access.sol
@@ -1,16 +1,16 @@
 contract c {
-    uint256[4] data;
+    uint[4] data;
 
-    function set(uint256 index, uint256 value) public returns (bool) {
+    function set(uint index, uint value) public returns (bool) {
         data[index] = value;
         return true;
     }
 
-    function get(uint256 index) public returns (uint256) {
+    function get(uint index) public returns (uint) {
         return data[index];
     }
 
-    function length() public returns (uint256) {
+    function length() public returns (uint) {
         return data.length;
     }
 }
@@ -20,10 +20,10 @@ contract c {
 // compileToEwasm: also
 // ----
 // length() -> 4
-// set(uint256,uint256): 3, 4 -> true
-// set(uint256,uint256): 4, 5 -> FAILURE, hex"4e487b71", 0x32
-// set(uint256,uint256): 400, 5 -> FAILURE, hex"4e487b71", 0x32
-// get(uint256): 3 -> 4
-// get(uint256): 4 -> FAILURE, hex"4e487b71", 0x32
-// get(uint256): 400 -> FAILURE, hex"4e487b71", 0x32
+// set(uint,uint): 3, 4 -> true
+// set(uint,uint): 4, 5 -> FAILURE, hex"4e487b71", 0x32
+// set(uint,uint): 400, 5 -> FAILURE, hex"4e487b71", 0x32
+// get(uint): 3 -> 4
+// get(uint): 4 -> FAILURE, hex"4e487b71", 0x32
+// get(uint): 400 -> FAILURE, hex"4e487b71", 0x32
 // length() -> 4
