+++ test/libsolidity/semanticTests/builtinFunctions/msg_sig.sol
@@ -1,5 +1,5 @@
 contract test {
-    function foo(uint256 a) public returns (bytes4 value) {
+    function foo(uint a) public returns (bytes4 value) {
         return msg.sig;
     }
 }
@@ -7,4 +7,4 @@ contract test {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// foo(uint256): 0x0 -> 0x2fbebd3800000000000000000000000000000000000000000000000000000000
+// foo(uint): 0x0 -> 0x2fbebd3800000000000000000000000000000000000000000000000000000000
