+++ test/libsolidity/semanticTests/builtinFunctions/msg_sig_after_internal_call_is_same.sol
@@ -3,7 +3,7 @@ contract test {
         return msg.sig;
     }
 
-    function foo(uint256 a) public returns (bytes4 value) {
+    function foo(uint a) public returns (bytes4 value) {
         return boo();
     }
 }
@@ -11,4 +11,4 @@ contract test {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// foo(uint256): 0x0 -> 0x2fbebd3800000000000000000000000000000000000000000000000000000000
+// foo(uint): 0x0 -> 0x2fbebd3800000000000000000000000000000000000000000000000000000000
