+++ test/libsolidity/semanticTests/cleanup/bool_conversion.sol
@@ -1,5 +1,5 @@
 contract C {
-    function f(bool _b) public returns (uint256) {
+    function f(bool _b) public returns (uint) {
         if (_b) return 1;
         else return 0;
     }
@@ -13,11 +13,11 @@ contract C {
 // ----
 // f(bool): 0x0 -> 0x0
 // f(bool): 0x1 -> 0x1
-// f(bool): 0x2 -> 0x1
-// f(bool): 0x3 -> 0x1
-// f(bool): 0xff -> 0x1
+// f(bool): 0x2 ->
+// f(bool): 0x3 ->
+// f(bool): 0xff ->
 // g(bool): 0x0 -> 0x0
 // g(bool): 0x1 -> 0x1
-// g(bool): 0x2 -> 0x1
-// g(bool): 0x3 -> 0x1
-// g(bool): 0xff -> 0x1
+// g(bool): 0x2 ->
+// g(bool): 0x3 ->
+// g(bool): 0xff ->
